datatype: widen ompi_datatype_match_size kind/lang to uint32_t
ompi_datatype_match_size() took its datakind and datalang arguments as
uint16_t, but the OMPI_DATATYPE_FLAG_DATA_* values it is called with live
at bits 21-24 (0x00200000 through 0x01800000). Passing them through a
uint16_t truncated every flag to zero, so the language and kind
comparisons inside never matched a real predefined type. This silently
broke the routine for its actual callers, MPI_Type_match_size in both the
C and Fortran bindings, which pass those flags directly.

Widen datakind and datalang to uint32_t in the declaration and the
definition so the full flag value survives, and widen the mirroring
uint16_t parameter in the ompi_datatype_bigcount test's check_match
helper, which reproduced the same truncation before the call.

+++ ompi/datatype/ompi_datatype.h
@@ -340,7 +340,7 @@ ompi_datatype_copy_content_same_ddt( const ompi_datatype_t* type, size_t count,
     return 0;
 }
 
-OMPI_DECLSPEC const ompi_datatype_t* ompi_datatype_match_size( size_t size, uint16_t datakind, uint16_t datalang );
+OMPI_DECLSPEC const ompi_datatype_t* ompi_datatype_match_size( size_t size, uint32_t datakind, uint32_t datalang );
 
 /*
  *
